django-api/requirements.txt
-i https://pypi.org/simple
asgiref==3.3.1
django-cors-headers==3.7.0
django==3.1.7
djangorestframework-jwt==1.11.0
djangorestframework==3.12.2
gunicorn==20.0.4
pyjwt==1.7.1
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6